requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.2.0

# Development Dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1